Rename context folder
