+++ runner/src/context-builder/context-builder.test.ts
@@ -3,7 +3,7 @@ import type fetch from 'node-fetch';
 import type DmlHandler from '../dml-handler';
 import IndexerConfig from '../indexer-config';
 import { LogLevel } from '../indexer-meta/log-entry';
-import ContextBuilder from './context';
+import ContextBuilder from './context-builder';
 
 describe('ContextBuilder unit tests', () => {
   const MOCK_CONFIG = {
