--- runner/src/context/index.ts
@@ -1,2 +1,0 @@
-export { default } from './context';
-export type { ContextObject } from './context';
