+++ runner/src/indexer/indexer.test.ts
@@ -7,8 +7,8 @@ import { LogLevel } from '../indexer-meta/log-entry';
 import IndexerConfig from '../indexer-config/indexer-config';
 import { IndexerStatus } from '../indexer-meta';
 import type IndexerMeta from '../indexer-meta';
-import ContextBuilder from '../context';
-import { type ContextObject } from '../context';
+import ContextBuilder from '../context-builder';
+import { type ContextObject } from '../context-builder';
 
 describe('Indexer unit tests', () => {
   const SIMPLE_REDIS_STREAM = 'test:stream';
